Renamed Posix and Windows timeout modules consistently with others.
Related to #1931. Assuming tests on CI pass, that issue is now done.
